+++ src/Persons/Supervisor.h
@@ -11,8 +11,8 @@ using namespace std;
 class Supervisor: public Employee {
 	vector<Employee *> teamvec;
 public:
-	Supervisor(string name, int age, unsigned long int phone, string email,
-			unsigned long int nif, unsigned long int wage);
+	Supervisor(string name, int age, unsigned int phone, string email,
+			unsigned int nif, unsigned int wage);
 	virtual ~Supervisor();
 	void addEmployee(Employee *e1);
 	bool removeEmployee(string name);
